feat(124258): :hammer: refatora solicitacoes codae
