+++ sme_sigpae_api/paineis_consolidados/urls.py
@@ -2,14 +2,13 @@ from django.urls import include, path
 from rest_framework import routers
 
 from .api import viewsets
+from .solicitacoes_codae.api.viewsets import CODAESolicitacoesViewSet
 
 router = routers.DefaultRouter()
 router.register(
     "solicitacoes-genericas", viewsets.SolicitacoesViewSet, "solicitacoes_genericas"
 )
-router.register(
-    "codae-solicitacoes", viewsets.CODAESolicitacoesViewSet, "codae_solicitacoes"
-)
+router.register("codae-solicitacoes", CODAESolicitacoesViewSet, "codae_solicitacoes")
 router.register(
     "nutrisupervisao-solicitacoes",
     viewsets.NutrisupervisaoSolicitacoesViewSet,
